backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Elasticsearch
elasticsearch[async]==8.11.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Background Tasks
celery[redis]==5.3.6
flower==2.0.1

# Email
boto3==1.34.14

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==23.12.1
isort==5.13.2
mypy==1.8.0
flake8==7.0.0

# Monitoring
sentry-sdk[fastapi]==1.39.1
prometheus-client==0.19.0
